refactored code to skip iterations where temp_index is similar to the end or index

+++ sort.h
@@ -23,7 +23,7 @@ void bubble_sort(int *array, size_t size);
 void selection_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void quick_sort(int *array, size_t size);
-size_t lomuto_part(int *array, size_t size, size_t start, size_t end);
+size_t lomuto_part(int *array, size_t size, int start, int end);
 void swap(int *first, int *second, const int *array, size_t size);
 void sort_algo(int *array, size_t size, int start, int end);
 #endif
